勤怠修正10/11

+++ demo2309/src/main/java/com/example/demo/repository/AttendanceCorrectionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.AttendanceCorrectionEntity;
 
-public interface AttendanceCorrectionRepository extends JpaRepository<AttendanceCorrectionEntity, Long> {
+public interface AttendanceCorrectionRepository extends JpaRepository<AttendanceCorrectionEntity, Integer> {
   
 }
